--- src/main/java/CloneProject/InstagramClone/InstagramService/domain/User.java
@@ -1,21 +1,0 @@
-package CloneProject.InstagramClone.InstagramService.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long uniqueId;
-    private String email;
-    private String password;
-    private String nickname;
-}
