+++ src/main/java/CloneProject/InstagramClone/InstagramService/dto/SignInDto.java
@@ -1,0 +1,13 @@
+package CloneProject.InstagramClone.InstagramService.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Data;
+
+@Data
+public class SignInDto {
+    @NotBlank
+    private String email;
+
+    @NotBlank
+    private String password;
+}
