+++ src/main/java/CloneProject/InstagramClone/InstagramService/dto/SignUpDto.java
@@ -1,0 +1,16 @@
+package CloneProject.InstagramClone.InstagramService.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Data;
+
+@Data
+public class SignUpDto {
+    @NotBlank
+    private String email;
+
+    @NotBlank
+    private String password;
+
+    @NotBlank
+    private String nickname;
+}
