+++ src/main/java/CloneProject/InstagramClone/InstagramService/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package CloneProject.InstagramClone.InstagramService.repository;
 
-import CloneProject.InstagramClone.InstagramService.domain.User;
+import CloneProject.InstagramClone.InstagramService.vo.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User,Long> {
