+++ src/main/java/CloneProject/InstagramClone/InstagramService/vo/Role.java
@@ -1,0 +1,25 @@
+package CloneProject.InstagramClone.InstagramService.vo;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+@Entity(name="roles")
+public class Role {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @NotBlank
+    @Column
+    private String name;
+
+    @ManyToMany(mappedBy = "roles")
+    private List<User> users;
+}
