+++ src/main/java/CloneProject/InstagramClone/InstagramService/vo/User.java
@@ -1,0 +1,36 @@
+package CloneProject.InstagramClone.InstagramService.vo;
+
+import jakarta.annotation.Nullable;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Getter
+@Setter
+@Entity(name="users")
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String email;
+
+    @Column(nullable = false)
+    private String password;
+
+    @Column(nullable = false)
+    private String nickname;
+
+    @JoinTable(
+            name="users_roles",
+            joinColumns = {@JoinColumn(name="USER_ID",referencedColumnName = "ID")},
+            inverseJoinColumns = {@JoinColumn(name="ROLE_ID",referencedColumnName = "ID")})
+    @ManyToMany(fetch = FetchType.EAGER,cascade = CascadeType.ALL)
+    private List<Role> roles = new ArrayList<>();
+}
